+++ tests/wasm/generic_import_ref.js
@@ -1,3 +1,7 @@
+exports.takeRefPrimitive = function(x) {
+  return x + 1;
+};
+
 exports.RefThing = class RefThing {
   constructor(val) {
     this.val = val;